requirements.txt
streamlit==1.35.0
transformers==4.41.1
torch==2.2.2
pinecone-client==4.1.0
langchain==0.2.1
langchain-community==0.2.1
python-dotenv==1.0.1
openai==1.30.3
tiktoken==0.7.0
streamlit_chat==0.1.1
unstructured